Updated the Signup system. The SignupEmail class will now adjust the database before sending off the email. This includes deleting old recovery codes tied to the email, and checking to see if the uniquely generated code already exists in the table as it is the primary key for it. This allows us to reuse the code incase we need to resend the Signup email again for email verification purposes.
Added the Account Recovery system. This will check to see if an account has a verfied email.
Should the email be verified, we will send a recovery email to the client with a unique 8 character
code. They can use this code to access the password updating system that is "outside" the login system,
meaning, they are not logged in and changing the password from the settings within.

ErrorHandler, I only updated a comment, nothing functional.

I tried to get .gitignore to ignore my log files, but it failed to do so. Only slightly improving with git.