Optimized error responses for signup. Each individual check will generate its own unique error code. Future updates will prevent front side from seeing some codes as only the server will need to be privy. Currently the front side does not interpret the code for the client, as it only send the response code in the URL query. These codes are 4 digits in length. One specific const "BACKSIDE_ONLY" carries the code number for which we can check against to prevent passing to front side visibility (everything greater than should not be shown). Front Side signup will now ask for first and last name which are optional.
Future update will need to send an error report for "BACKSIDE_ONLY" codes.

The "check" class will now return a bool for key_check() to assist in abbreviating flow for classes that extend it.
check class now has a set of its own error codes that are recognizable with its 3 digit format.

The "email" class now has a property that stores the bool response from mail(). This assists in the flow and
error checking for the signup system.

autoloader will now find the correct directory for the signup system.

start.php will no longer attempt to update a location of a client who isnt logged in.