Optimized error responses for login. Each individual check will generate its own unique error code.
Currently the front side does not interpret the code for the client, as it only send the response code in the
URL query. These codes are 4 digits in length. One specific const "BACKSIDE_ONLY" carries the code number
for which we can check against to prevent passing to front side visibility (everything greater than should not be shown).

Errors generated by the server will report error to log. Client errors will not report. No client input will be sent back
to the front side page to preload in the inputs.